GH-0 fix run and documents
