+++ db/scripts/wait-for.sh
@@ -1,6 +1,4 @@
-#!/bin/bash
-# wait-for-postgres.sh
-
+#!/bin/sh
 set -e
 
 host="$1"
